backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
python-dotenv==1.0.1
openai==1.58.1
pydantic==2.10.4
sse-starlette==2.2.1
